% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{anno_info}
\alias{anno_info}
\title{Annotation information from Pvirgatum_516_v5.1.annotation_info}
\format{
A data frame with 80278 rows and 16 variables.
}
\usage{
data(anno_info)
}
\description{
A dataset containing the v5.1 annotation information for Panicum virgatum.
The variables are as follows:
}
\details{
\itemize{
\item pacID ID of the Panicum virgatum gene from the annotation file.
\item locusName The gene name from Panicum virgatum. Only distinct locus
names are included in this dataset.
\item transcriptName The transcript name from P. virgatum. Only the first
transcript variant is included in this dataset.
\item peptideName The protein name for the variant. Only the first protein
variant is included in this dataset.
\item Pfam
\item Panther Panther categories associated with this locus.
\item KOG assignments from the euKaryotic Orthologous Groups (KOG) database
\item ec EC numbers in Enzyme Nomenclature.
\item KO KEGG Orthology (KO) numbers associated with this locusName.
\item GO Gene ontology entries associated with this locusName.
\item Best-hit-arabi-name Best Arabidopsis thaliana homolog locus name.
\item arabi-symbol Best A. thaliana homolog gene name.
\item arabi-defline Best A. thaliana homolog gene annotation.
\item Best-hit-rice-name Best rice homolog locus name.
\item rice-symbol Best rice homolog gene name.
\item rice-defline Best rice homolog gene annotation.
}
}
\references{
\url{https://www.nature.com/articles/s41586-020-03127-1}
}
\author{
Alice MacQueen \email{alice.macqueen@utexas.edu}
}
\keyword{data}
