% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pvdiv_gwas.R
\name{asv_best_PC_df}
\alias{asv_best_PC_df}
\title{Return best number of PCs in terms of lambda_GC following Cattrell's rule.}
\usage{
asv_best_PC_df(df)
}
\arguments{
\item{df}{Dataframe of phenotypes where the first column is NumPCs and
subsequent column contains lambda_GC values for some phenotype.}
}
\value{
A dataframe containing the best lambda_GC value and number of PCs
for each phenotype in the data frame.
}
\description{
Given a dataframe created using pvdiv_lambda_GC, this function
returns the lambda_GC that is closest to 1 for each column in the
dataframe.
}
