% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pvdiv_gwas.R
\name{dots}
\alias{dots}
\alias{advanced...}
\title{Set an advanced argument}
\usage{
dots(name, value, ...)
}
\arguments{
\item{name}{Name of the advanced argument to look for in \code{...}}

\item{value}{The default value of the advanged argument. If this advanced
argument is used in several of your functions, we recommend using
\code{getOption('value')} and explaining this option in your package
vignette geared towards experienced users.}

\item{...}{Advanced arguments. See \link[methods]{dotsMethods}.}
}
\description{
This function is taken from the lcolladotor/dots package. It is
the only function used from this package and is added to this package
wholesale to reduce user installation burden. Please use the original
package from Github if you use this in your own work.
}
\details{
From the \code{...} argument used in your function, find if a specific
argument was included and extract its value.

Note that you can make dots() even more powerful by using \link{getOption}
to define \code{value}. This is particularly useful if you use the
same advanced argument in several functions.
}
\examples{

## Simple example that calculates the max between 'x' and 'y' with a
## specified minimum value to return.

minMax <- function(x, y, ...) {
    minValue <- dots('minValue', 0, ...)
    res <- max(x, y, minValue)
    return(res)
}
minMax(1:2, 3:4)
minMax(1:2, 3:4, minValue = 5)


## Arguably these examples are simple, but the idea is that dots()
## can simplify very long function calls where some parameters will be used
## by a minority of the users.

}
\seealso{
\link[methods]{dotsMethods}
}
\author{
L. Collado-Torres
}
