% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pvdiv_mash_results.R
\name{get_estimated_pi}
\alias{get_estimated_pi}
\title{Return the estimated mixture proportions. Use get_estimated_pi to
extract the estimates of the mixture proportions for different types of
covariance matrix. This tells you which covariance matrices have most of
the mass.}
\usage{
get_estimated_pi(m, dimension = c("cov", "grid", "all"))
}
\arguments{
\item{m}{the mash result}

\item{dimension}{indicates whether you want the mixture proportions for the
covariances, grid, or all}
}
\value{
a named vector containing the estimated mixture proportions.
}
\description{
Return the estimated mixture proportions. Use get_estimated_pi to
extract the estimates of the mixture proportions for different types of
covariance matrix. This tells you which covariance matrices have most of
the mass.
}
\details{
If the fit was done with \code{usepointmass=TRUE} then the first
element of the returned vector will correspond to the null, and the
remaining elements to the non-null covariance matrices. Suppose the fit
was done with $K$ covariances and a grid of length $L$. If
\code{dimension=cov} then the returned vector will be of length $K$
(or $K+1$ if \code{usepointmass=TRUE}).  If \code{dimension=grid} then
the returned vector will be of length $L$ (or $L+1$).  If
\code{dimension=all} then the returned vector will be of length $LK$ (or
$LK+1$). The names of the vector will be informative for which
combination each element corresponds to.
}
