% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pvdiv_gwas_results.R
\name{get_fbm_upset_df}
\alias{get_fbm_upset_df}
\title{Plot the -log10p-value distributions for two univariate GWAS against
one another using ggplot.}
\usage{
get_fbm_upset_df(effects, snp, metadata, thr = 7, quantile = NA)
}
\arguments{
\item{effects}{fbm created using 'pvdiv_standard_gwas'.
Saved under the name "gwas_effects_{suffix}.rds" and can be loaded into
R using the bigstatsr function "big_attach".}

\item{snp}{A "bigSNP" object; load with \code{bigsnpr::snp_attach()}.
Here, genomic information for Panicum virgatum.}

\item{metadata}{Metadata created using 'pvdiv_standard_gwas'.
Saved under the name "gwas_effects_{suffix}_associated_metadata.csv"}

\item{thr}{Numeric. Threshold above which SNP/row is kept for comparisons.}

\item{quantile}{Numeric. Top quantile/percentile to keep for each GWAS for
comparisons.}
}
\description{
This function takes GWAS results from the dive_ functions of
snpdiver that create FBM of univariate GWAS effects. It creates a
dataframe from these results suitable for an Upset plot, containing only
the rows/SNPs significant in at least one univariate GWAS at the
-log10p threshold specified.
}
