% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pvdiv_mash_results.R
\name{get_GxE}
\alias{get_GxE}
\title{Get dataframes of types of GxE from mash}
\usage{
get_GxE(m, factor = 0.4, thresh = 0.05, window_bp = 20000)
}
\arguments{
\item{m}{An object of type mash}

\item{factor}{a number between 0 and 1. The factor within which effects are
considered to be shared.}

\item{thresh}{Numeric. The threshold for including an effect in the assessment}

\item{window_bp}{The window size, in base pairs, to thin to one marker
within.}
}
\description{
Get dataframes of types of GxE from mash
}
