% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pvdiv_mash_results.R
\name{get_mash_metadata}
\alias{get_mash_metadata}
\title{Get switchgrass metadata for SNPs in a mash object.}
\usage{
get_mash_metadata(m, snp, suffix = "", saveoutput = FALSE)
}
\arguments{
\item{m}{A mash object (outputted by mash).}

\item{snp}{A bigsnp object}

\item{suffix}{Character. Optional. A unique suffix used to save the files,
instead of the current date & time.}

\item{saveoutput}{Logical. Should the output be saved to the path?}
}
\value{
A data frame containing details for the alternate allele of each SNP
such as its frequency in the bigsnp object, the average latitude,
longitude, elevation, and year that the SNP was sampled, and how
frequently individuals with the SNP have specific subpop and ecotype
identities in the bigsnp object.
}
\description{
Takes a mash object and a bigsnp object and returns
a dataframe containing summary metadata for each SNP in the mash object.
This information includes details for the alternate allele of each SNP
such as its frequency in the bigsnp object, the average latitude,
longitude, elevation, and year that the SNP was sampled, and how
frequently individuals with the SNP have specific subpop and ecotype
identities in the bigsnp object.
NB: This is VERY slow and best for up to tens of thousands of
SNPs, NOT millions of SNPs.
}
\note{
This function is VERY slow and best for up to tens of thousands of
SNPs, NOT millions of SNPs. Use vcftools instead for millions of SNPs.
}
