% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{gff_gene}
\alias{gff_gene}
\title{GFF3 Gene information}
\format{A data frame with 1236881 rows and 9 variables}
\usage{
data(gff_gene)
}
\description{
A dataset containing the gene information in GFF3 format.
The variables are as follows:
}
\details{
\itemize{
\item seqid. name of the chromosome or scaffold; here, of Panicum virgatum.
\item source. name of the program that generated this feature.
\item type. type of feature. Term or accession from the SOFA sequence ontology.
\item start. Start position of the feature, with sequence numbering starting at 1.
\item end. End position of the feature, with sequence numbering starting at 1.
\item score. A floating point value.
\item strand. defined as + (forward) or - (reverse).
\item phase. One of '0', '1' or '2'. '0' indicates that the first base of the feature is the first base of a codon, '1' that the second base is the first base of a codon, and so on.
\item attributes. A semicolon-separated list of tag-value pairs, providing additional information about each feature. Some of these tags are predefined, e.g. ID, Name, Alias, Parent.
}
}
\references{
\url{data_blah.com}
}
\author{
Alice MacQueen \email{alice.macqueen@utexas.edu}
}
\keyword{data}
