% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{gff_metadata}
\alias{gff_metadata}
\title{Metadata lines for the gene GFF3}
\format{
A data frame with 3 rows and 1 variable.
}
\usage{
data(gff_metadata)
}
\description{
A dataset containing the three metadata lines for the gff3 file gff_gene.
}
\references{
\url{data_blah.com}
}
\author{
Alice MacQueen \email{alice.macqueen@utexas.edu}
}
\keyword{data}
