% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pvdiv_mash_results.R
\name{mash_thin_named_markers}
\alias{mash_thin_named_markers}
\title{Thin significant markers using names of the markers}
\usage{
mash_thin_named_markers(m, markers, window_bp = 20000, thresh = 0.05)
}
\arguments{
\item{m}{An object of type mash}

\item{markers}{A vector of marker names also found in the mash object.}

\item{window_bp}{Numeric. The window size in base pairs, within which to
keep one significant SNP.}

\item{thresh}{What is the threshold to call an effect significant? Default is
0.05.}
}
\value{
A tbl_df of Marker, Chr, Pos, and the minimum local false sign rate
(combined across all conditions for plotting purposes using the
Bonferroni method, aka, multiplying the minimum p-value by the number
of conditions).
}
\description{
Thin significant markers using names of the markers
}
