% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pvdiv_gwas.R
\name{mt.rawp2adjp}
\alias{mt.rawp2adjp}
\title{Adjusted p-values for simple multiple testing procedures}
\usage{
mt.rawp2adjp(rawp, proc=c("Bonferroni", "Holm", "Hochberg", "SidakSS",
    "SidakSD", "BH", "BY","ABH","TSBH"), alpha = 0.05, na.rm = FALSE)
}
\arguments{
\item{rawp}{A vector of raw (unadjusted) p-values for each hypothesis under
consideration. These could be nominal p-values, for example, from
t-tables, or permutation p-values as given in mt.maxT and mt.minP. If
the mt.maxT or mt.minP functions are used, raw p-values should be given
in the original data order, ordered by the index of that data.}

\item{proc}{A vector of character strings containing the names of the
multiple testing procedures for which adjusted p-values are to be
computed. This vector should include any of the following: "Bonferroni",
"Holm", "Hochberg", "SidakSS", "SidakSD", "BH", "BY", "ABH", "TSBH".}

\item{alpha}{A nominal type I error rate, or a vector of error rates, used
for estimating the number of true null hypotheses in the two-stage
Benjamini & Hochberg procedure ("TSBH"). Default is 0.05.}

\item{na.rm}{An option for handling NA values in a list of raw p-values.
If FALSE, the number of hypotheses considered is the length of the vector
of raw p-values. Otherwise, if TRUE, the number of hypotheses is the
number of raw p-values which were not NAs.}
}
\value{
A list with components: adjp, index, h0.ABH, h0.TSBH. See multtest
package on Bioconductor for details.
}
\description{
This function computes adjusted p-values for simple multiple
testing procedures from a vector of raw (unadjusted) p-values. The
procedures include the Bonferroni, Holm (1979), Hochberg (1988), and
Sidak procedures for strong control of the family-wise Type I error
rate (FWER), and the Benjamini & Hochberg (1995) and Benjamini &
Yekutieli (2001) procedures for (strong) control of the false discovery
rate (FDR). The less conservative adaptive Benjamini & Hochberg (2000)
and two-stage Benjamini & Hochberg (2006) FDR-controlling procedures are
also included. This function is taken from the multtest package. It is
the only function used from this package and is added to this package
wholesale to reduce user installation burden.
}
\author{
Sandrine Dudoit, http://www.stat.berkeley.edu/~sandrine,

Yongchao Ge, yongchao.ge@mssm.edu,

Houston Gilbert, http://www.stat.berkeley.edu/~houston.
}
