% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{phenotypes}
\alias{phenotypes}
\title{Phenotypes available for the Panicum virgatum diversity panel.}
\format{A data frame with 785 rows and 4 variables}
\usage{
data(phenotypes)
}
\description{
A dataset containing phenotypes for the Panicum virgatum diversity panel.
The variables are as follows:
}
\details{
\itemize{
\item PLANT_ID. ID of the Panicum virgatum genotype sequenced in the diversity panel.
\item GWAS_CT. The number of times this genotype was cloned and planted in the ten 2018 sites for the Panicum virgatum diversity panel.
\item BRKG_TC_EOS_2018.
\item BRKG_DEAD_2018.
}
}
\references{
\url{data_blah.com}
}
\author{
Alice MacQueen \email{alice.macqueen@utexas.edu}
}
\keyword{data}
