% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pvdiv_gwas_results.R
\name{pvdiv_anno_subset}
\alias{pvdiv_anno_subset}
\title{Make a SNP dataframe subset from annotation table output.}
\usage{
pvdiv_anno_subset(anno_df, snp)
}
\arguments{
\item{anno_df}{One dataframe of annotations from pvdiv_table_topsnps(). This
dataframe needs to contain the columns CHR and region_start. It's
recommended that you set rangevector = 0 in pvdiv_table_topsnps() to get
the SNP itself using this function.}

\item{snp}{A \code{FBM.code256} object. Genomic information for Panicum virgatum.
Contact tjuenger <at> utexas <dot> edu to obtain this information
pre-publication.}
}
\value{
A \code{tbl_df()} of the SNP calls for all individuals for the
subset of SNPs in the annotation data frame.
}
\description{
Given a dataframe created using pvdiv_table_topsnps(), this
function creates a dataframe of SNP calls for the subset of SNPs from
this annotation table.
}
\note{
bigsnpr already has functions to subset its SNP file format and return
a bed file and to return its own SNP file format. This function is useful
if you instead want a small(ish) dataframe of SNPs that can be
manipulated using data frame tools in R.
}
