% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pvdiv_gwas_results.R
\name{pvdiv_bigsnp2tibble}
\alias{pvdiv_bigsnp2tibble}
\title{Convert A \code{FBM.code256} SNP subset to a dataframe.}
\usage{
pvdiv_bigsnp2tibble(snp)
}
\arguments{
\item{snp}{A \code{FBM.code256} object. Genomic information for Panicum virgatum.
SNP data is available at doi:10.18738/T8/ET9UAU}
}
\value{
A \code{tbl_df()} of the SNP calls for all individuals for the
subset of SNPs on that chromosome between the two positions specified.
}
\description{
This function creates a dataframe of SNP calls for a
\code{FBM.code256} object. We recommend that this be a dataframe for a small
region.
}
\note{
bigsnpr already has functions to subset its SNP file format and return
a bed file and to return its own SNP file format. This function is useful
if you instead want a small(ish) dataframe of SNPs that can be
manipulated using data frame tools in R.
}
