% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pvdiv_gwas_results.R
\name{pvdiv_bigsnp_subset}
\alias{pvdiv_bigsnp_subset}
\title{Make a SNP dataframe subset from annotation table or a genomic range.}
\usage{
pvdiv_bigsnp_subset(snp, type = c("anno", "range"), anno_df, chr, pos1, pos2)
}
\arguments{
\item{snp}{A \code{FBM.code256} object. Genomic information for Panicum virgatum.
Contact tjuenger <at> utexas <dot> edu to obtain this information
pre-publication.}

\item{type}{One of "anno" or "range", depending on if you are using an
annotation dataframe from pvdiv_table_topsnps() or a genomic interval.}

\item{anno_df}{One dataframe of annotations from pvdiv_table_topsnps(). This
dataframe needs to contain the columns CHR and region_start. It's
recommended that you set rangevector = 0 in pvdiv_table_topsnps() to get
the SNP itself using this function.}

\item{chr}{Character string. The chromsome (e.g., "Chr01K") to get SNPs from.}

\item{pos1}{Integer. The low position to start getting SNPs from.}

\item{pos2}{Integer. The high position to stop getting SNPs from.}
}
\value{
A \code{FBM.code256} object for the subset of SNPs in the annotation data
frame or in the genomic interval.
}
\description{
Given a dataframe created using pvdiv_table_topsnps(), this
function creates a dataframe of SNP calls for the subset of SNPs from
this annotation table.
}
\note{
This function is a wrapper around bigsnpr functions to subset its SNP
file format that may be useful if you have a small interval to look at or
a small number of SNPs from an annotation table.
}
