% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pvdiv_gwas_results.R
\name{pvdiv_fdr}
\alias{pvdiv_fdr}
\title{Compute Adjusted False Discovery Rate for pvdiv_gwas output.}
\usage{
pvdiv_fdr(gwas_obj, alpha = 0.05, proc = "BH")
}
\arguments{
\item{gwas_obj}{A dataframe with a \code{p.value} column.}

\item{alpha}{The type I error threshold. Default is 0.05.}

\item{proc}{The procedure (used wholesale from \code{multtest::mt.rawp2adjp})
used to calculate the False Discovery Rate. Default is Benjamini
Hochberg, "BH".}
}
\value{
A \code{tbl_df()} of the original dataframe with a new column named
\code{FDR_adjusted_p}.
}
\description{
Determines the false discovery rate for p-values from a
dataframe. By default, this function uses a Benjamini-Hochberg
correction on a column called "p.value" to determine a new column of
FDR-adjusted p-values.
}
\examples{
\dontrun{
gwas_df <- read_csv(file = "LEN.full.GWAS.csv.gz", col_names = TRUE,
    col_types = "ciinnnn")
gwas_fdr_1 <- pvdiv_fdr(gwas_obj = gwas_df, alpha = 0.1)
gwas_fdr_05 <- pvdiv_fdr(gwas_obj = gwas_df)
gwas_fdr_bonf <- pvdiv_fdr(gwas_obj = gwas_df, proc = "Bonferroni")
}

}
