% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pvdiv_standard_gwas.R
\name{pvdiv_kinship}
\alias{pvdiv_kinship}
\title{Find a kinship matrix using the van Raden method.}
\usage{
pvdiv_kinship(snp, ind.row = NA, saveoutput = FALSE)
}
\arguments{
\item{snp}{A "bigSNP" object; load with bigsnpr::snp_attach().}

\item{ind.row}{An integer vector of the rows (individuals) to find a
kinship matrix for. Defaults to all rows.}

\item{saveoutput}{Logical. Should the output be saved to the working
directory?}
}
\value{
A kinship matrix with labeled rows and columns.
}
\description{
Calculate the kinship matrix using code from GAPIT and bigsnpr
and the methods of VanRaden (2009, J. Dairy Sci. 91:4414???C4423). Note
that this matrix cannot currently be used with the GWAS methods in
bigsnpr; however, this matrix could be used with other GWAS programs.
}
\examples{
\dontrun{
K <- pvdiv_kinship(snp = snp, saveoutput = TRUE)
}

}
