% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pvdiv_gwas.R
\name{pvdiv_manhattan}
\alias{pvdiv_manhattan}
\title{Create a Manhattan plot with ggplot2.}
\usage{
pvdiv_manhattan(effects, ind = NULL, snp = NULL, thr = NULL, ncores = 1)
}
\arguments{
\item{effects}{Either a gwas effects RDS or FBM object created using
'pvdiv_standard_gwas' (with savetype = "rds" or savetype = "fbm"). If a
fbm, this file is saved under the name "gwas_effects_{suffix}.rds" and
can be loaded into R using the bigstatsr function "big_attach".}

\item{ind}{If effects is a FBM object, this should be the row number of the
phenotype from the associated metadata for the FBM object.}

\item{snp}{If effects is a FBM object, you must also supply a "bigSNP" object;
load into R with \code{bigsnpr::snp_attach()}.}

\item{thr}{Numeric. Significance threshold plotted as a horizontal line.
Default is Bonferroni.}

\item{ncores}{Integer. Number of cores to use for parallelization.}

\item{tol}{Numeric. Tolerance for optional Genomic Control coefficient.}
}
\value{
A ggplot2 plot.
}
\description{

}
