% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{pvdiv_phenotypes}
\alias{pvdiv_phenotypes}
\title{Phenotypes available for the Panicum virgatum diversity panel.}
\format{
A data frame with 1114 rows and 13 variables
}
\usage{
data(pvdiv_phenotypes)
}
\description{
A dataset containing phenotypes for the Panicum virgatum diversity panel.
The variables are as follows:
}
\details{
\itemize{
\item PLANT_ID. ID of the Panicum virgatum genotype sequenced in the
diversity panel.
\item GWAS_CT. The number of times that genotype was clonally propagated
and planted at these common gardens.
\item MAT. Mean annual temperature at the genotype's location of origin.
\item bio17. Precipitation in the driest quarter at the genotype's location
of origin.
\item bio4. Temperature seasonality at the genotype's location of origin.
\item bio16. Precipitation in the wettest quarter at the genotype's
location of origin.
\item AHM. Annual heat moisture index, calculated as (MAT+10)/(MAP/1000),
at the genotype's location of origin.
\item bio2. Mean diurnal temperature range at the genotype's location of
origin.
\item bio5. Max temperature of the warmest month at the genotype's location
of origin.
\item FRAC_SRV_THREE. Fraction of individuals surviving winter 2018/2019
at three northern gardens (BRKG, LINC, CLMB).
\item CLMB_BIOMASS. Biomass at the end of the 2019 season in Columbia,
Missouri.
\item KBSM_BIOMASS. Biomass at the end of the 2019 season in Hickory
Corners, Michigan (Kellog Biological Station).
\item PKLE_BIOMASS. Biomass at the end of the 2019 season in Austin, Texas
(Pickle Research Campus).
}
}
\references{
\url{https://www.nature.com/articles/s41586-020-03127-1}
}
\author{
Alice MacQueen \email{alice.macqueen@utexas.edu}
}
\keyword{data}
