% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pvdiv_gwas_results.R
\name{pvdiv_range_subset}
\alias{pvdiv_range_subset}
\title{Make a SNP dataframe subset for a region of the genome.}
\usage{
pvdiv_range_subset(snp, chr, pos1, pos2)
}
\arguments{
\item{snp}{A \code{FBM.code256} object. Genomic information for Panicum virgatum.
Contact tjuenger <at> utexas <dot> edu to obtain this information
pre-publication.}

\item{chr}{Character string. The chromsome (e.g., "Chr01K") to get SNPs from.}

\item{pos1}{Integer. The low position to start getting SNPs from.}

\item{pos2}{Integer. The high position to stop getting SNPs from.}
}
\value{
A \code{tbl_df()} of the SNP calls for all individuals for the
subset of SNPs on that chromosome between the two positions specified.
}
\description{
This function creates a dataframe of SNP calls for a subset of
SNPs from one region on one chromosome.
}
\note{
bigsnpr already has functions to subset its SNP file format and return
a bed file and to return its own SNP file format. This function is useful
if you instead want a small(ish) dataframe of SNPs that can be
manipulated using data frame tools in R.
}
