% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pvdiv_bigsnp2mashr.R
\name{pvdiv_results_in_folder}
\alias{pvdiv_results_in_folder}
\title{Identify phenotype names from bigsnpr results in a folder.}
\usage{
pvdiv_results_in_folder(path = ".", pattern = "*.rds")
}
\arguments{
\item{path}{File path to the files from bigsnpr, a character string.
Defaults to the current working directory.}

\item{pattern}{Pattern within the filename to match. Default is "*.rds".}
}
\value{
A vector of phenotype names.
}
\description{
Creates a vector of phenotype names from bigsnpr results.
}
\examples{
\dontrun{gwas_results_in_folder(path = system.file("inst/extdata",
package = "switchgrassGWAS"))}
\dontrun{gwas_results_in_folder(path = "path/to/gwas/results")}

}
