% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pvdiv_standard_gwas.R
\name{pvdiv_standard_gwas}
\alias{pvdiv_standard_gwas}
\title{Wrapper for Juenger lab standard GWAS functions for switchgrass.}
\usage{
pvdiv_standard_gwas(snp, df = switchgrassGWAS::phenotypes,
  type = c("linear", "logistic"), covar = NULL, ncores = 1,
  lambdagc = TRUE, outputdir = ".", savegwas = FALSE,
  saveplots = TRUE, saveannos = FALSE, txdb = NULL, ...)
}
\arguments{
\item{snp}{A "bigSNP" object; load with bigsnpr::snp_attach(). Here, genomic
information for Panicum virgatum. Contact tjuenger <at> utexas <dot> edu
to obtain this information pre-publication.}

\item{df}{Dataframe of phenotypes where the first column is PLANT_ID.}

\item{type}{Character string. Type of univarate regression to run for GWAS.
Options are "linear" or "logistic".}

\item{covar}{Optional covariance matrix to include in the regression. You
can generate these using \code{pvdiv_autoSVD()}.}

\item{ncores}{Number of cores to use. Default is one.}

\item{lambdagc}{Default is TRUE - should lambda_GC be used to find the best
population structure correction? Alternatively, you can provide a data
frame containing "NumPCs" and the phenotype names containing lambda_GC
values. This is saved to the output directory by pvdiv_standard_gwas and
saved or generated by pvdiv_lambda_GC.}

\item{outputdir}{String or file.path() to the output directory. Default is
the working directory.}

\item{savegwas}{Logical. Should the gwas output be saved as a rds to the
working directory? These files are typically quite large. Default is
FALSE.}

\item{saveplots}{Logical. Should Manhattan and QQ-plots be generated and
saved to the working directory? Default is TRUE.}

\item{saveannos}{Logical. Should annotation tables for top SNPs be generated
and saved to the working directory? Default is FALSE. Can take
additional arguments; requires a txdb.sqlite object used in
AnnotationDbi.}

\item{txdb}{A txdb object such as 'Pvirgatum_516_v5.1.gene.txdb.sqlite'.
Load this into your environment with AnnotationDbi::loadDb.}

\item{...}{Other arguments to \code{\link{pvdiv_lambda_GC}} or
\code{\link{pvdiv_table_topsnps}}.}
}
\value{
A big_SVD object.
}
\description{
This is a wrapper to
}
\examples{
\dontrun{
pvdiv_standard_gwas(snp, df = phenotypes, type = "linear", covar = svd,
    ncores = nb_cores(), lambdagc = TRUE, savegwas = TRUE, saveplots = TRUE,
    saveannos = TRUE, txdb = txdb)
}

}
