% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pvdiv_popgenome.R
\name{pvdiv_subset_popgenome}
\alias{pvdiv_subset_popgenome}
\title{Get GFF subsets}
\usage{
pvdiv_subset_popgenome(path, filename, seqid, start, end)
}
\arguments{
\item{path}{file.path() or character string. Path to the VCF file subset.}

\item{filename}{Character string. VCF subset filename}

\item{seqid}{Character string. name of the chromosome or scaffold; here, of Panicum virgatum.}

\item{start}{Integer. Start position of the feature, with sequence numbering starting at 1.}

\item{end}{Integer. End position of the feature, with sequence numbering starting at 1.}
}
\description{
Prepare a subset of the P. virgatum VCF file to be used in the
PopGenome R package. Given a VCF subset (created using VCFtools or some
other commandline program), this function will make a GFF and VCF folder
in the path, if they do not already exist, and place the VCF subset in
the correct format in the VCF folder, and the GFF subset in the correct
format in the GFF folder. These files can then be read into the R
package PopGenome.
}
