% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pvdiv_bigsnp2mashr.R
\name{pvdiv_top_effects_log10p}
\alias{pvdiv_top_effects_log10p}
\title{Step One of bigsnp2mashr}
\usage{
pvdiv_top_effects_log10p(path, gwas_rds, phenotype, numSNPs,
  model = c("linear", "logistic"), keep_score = FALSE, markers)
}
\arguments{
\item{path}{Path}

\item{gwas_rds}{RDS file with gwas results}

\item{phenotype}{Character vector. Single phenotype name}

\item{numSNPs}{Integer. Number of top SNPs to choose.}

\item{model}{One of linear or logistic. Type of GWAS model.}

\item{keep_score}{Logical. Keep p-values for clumping?}

\item{markers}{Marker names for the GWAS you ran}
}
\description{
Step One of bigsnp2mashr
}
