% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pvdiv_popgenome.R
\name{pvdiv_vcf_popgenome}
\alias{pvdiv_vcf_popgenome}
\title{Format the VCF subset}
\usage{
pvdiv_vcf_popgenome(path, filename)
}
\arguments{
\item{path}{file.path() or character string. Path to the VCF file subset.}

\item{filename}{Character string. VCF subset filename}
}
\description{
Format the VCF subset
}
