% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pvdiv_bigsnp2mashr.R
\name{s_hat_bigsnp}
\alias{s_hat_bigsnp}
\title{Step Two of bigsnp2mashr}
\usage{
s_hat_bigsnp(path, gwas_rds, phenotype, top_set, random_sample,
  markers = switchgrassGWAS::markers,
  markers2 = switchgrassGWAS::markers2, model = c("linear",
  "logistic"), clumped = FALSE)
}
\arguments{
\item{path}{Path}

\item{gwas_rds}{RDS file with gwas results}

\item{phenotype}{Character vector. Single phenotype name}

\item{top_set}{Top markers chosen}

\item{random_sample}{Numeric vector. Random sample of SNPs}

\item{markers}{Marker names for the GWAS you ran}

\item{markers2}{Markers to include if SNPs are LD clumped.}

\item{model}{One of linear or logistic. Type of GWAS model.}

\item{clumped}{Logical. Clump SNPs?}
}
\description{
Step Two of bigsnp2mashr
}
