% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pvdiv_gwas_results.R
\name{thresholdFBM}
\alias{thresholdFBM}
\title{Threshold a FBM}
\usage{
thresholdFBM(X, ind, thr, quantile = NA)
}
\arguments{
\item{X}{A FBM.}

\item{ind}{Numeric. Single column to threshold.}

\item{thr}{Numeric. Threshold above which row value is changed to "TRUE".}

\item{quantile}{Numeric. Top quantile/percentile to keep for each GWAS for
comparisons.}
}
\description{
Threshold a FBM
}
