% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pvdiv_standard_gwas.R
\name{pvdiv_standard_gwas}
\alias{pvdiv_standard_gwas}
\title{Juenger lab standard GWAS function.}
\usage{
pvdiv_standard_gwas(
  snp,
  df = switchgrassGWAS::pvdiv_phenotypes,
  type = c("linear", "logistic"),
  ncores = nb_cores(),
  outputdir = ".",
  covar = NULL,
  lambdagc = TRUE,
  savegwas = FALSE,
  savetype = c("rds", "fbm", "both"),
  suffix = "",
  saveplots = TRUE,
  saveannos = FALSE,
  txdb = NULL,
  minphe = 200,
  ...
)
}
\arguments{
\item{snp}{A "bigSNP" object; load with \code{bigsnpr::snp_attach()}.
Here, genomic information for Panicum virgatum. SNP data
is available at doi:10.18738/T8/ET9UAU}

\item{df}{Dataframe of phenotypes where the first column is PLANT_ID.}

\item{type}{Character string. Type of univarate regression to run for GWAS.
Options are "linear" or "logistic".}

\item{ncores}{Number of cores to use. Default is one.}

\item{outputdir}{String or file.path() to the output directory. Default is
the working directory.}

\item{covar}{Optional covariance matrix to include in the regression. You
can generate these using \code{pvdiv_autoSVD()}.}

\item{lambdagc}{Default is TRUE - should lambda_GC be used to find the best
population structure correction? Alternatively, you can provide a data
frame containing "NumPCs" and the phenotype names containing lambda_GC
values. This is saved to the output directory by pvdiv_standard_gwas and
saved or generated by pvdiv_lambda_GC.}

\item{savegwas}{Logical. Should the gwas output be saved to the
working directory? These files are typically quite large. Default is
FALSE.}

\item{savetype}{Character string. Type of GWAS save file. Options are 'rds',
which saves individual rds files for each GWAS; 'fbm', which saves one
filebacked big matrix (using the bigsnpr package), or 'both', which
saves both file types. These files are typically quite large.}

\item{suffix}{Optional character vector to give saved files a unique search
string/name.}

\item{saveplots}{Logical. Should Manhattan and QQ-plots be generated and
saved to the working directory? Default is TRUE.}

\item{saveannos}{Logical. Should annotation tables for top SNPs be generated
and saved to the working directory? Default is FALSE. Can take
additional arguments; requires a txdb.sqlite object used in
AnnotationDbi.}

\item{txdb}{A txdb object such as 'Pvirgatum_516_v5.1.gene.txdb.sqlite'.
Load this into your environment with AnnotationDbi::loadDb.}

\item{minphe}{Integer. What's the minimum number of phenotyped individuals
to conduct a GWAS on? Default is 200. Use lower values with caution.}

\item{...}{Other arguments to \code{\link{pvdiv_lambda_GC}} or
\code{\link{pvdiv_table_topsnps}}.}
}
\value{
A big_SVD object.
}
\description{
This function is a wrapper around the standard GWAS procedures
in the Juenger lab. Singular value decomposition of the SNPs is done to
get principal components for population structure correction; the 'best'
number of PCs is chosen as the one that makes lambda_GC, the Genomic
Control coefficient, closest to 1. (See the \code{lambdagc} parameter to set
this yourself.) Next, genome-wide association is conducted, and the GWAS
output can be saved, as well as Manhattan plots, QQ-plots, and annotation
information for the top SNPs for each phenotype.
}
\examples{
\dontrun{
# Here we specify that we do want to generate and save the gwas dataframes,
# the Manhattan and QQ-plots, and the annotation tables.
pvdiv_standard_gwas(snp, df = pvdiv_phenotypes, type = "linear", covar = svd,
    ncores = nb_cores(), lambdagc = TRUE, savegwas = TRUE, saveplots = TRUE,
    saveannos = TRUE, txdb = txdb)
}
# In this example, we run GWAS on all the phenotypes in pvdiv_phenotypes
# using an example SNP set of ~1800 SNPs.
snpfile <- system.file("extdata", "example_bigsnp.rds", package = "switchgrassGWAS")
library(bigsnpr)
snp <- snp_attach(snpfile)
pvdiv_standard_gwas(snp, df = pvdiv_phenotypes, type = "linear", savegwas = FALSE,
    saveplots = FALSE, ncores = 1)


}
